move path conversion to support.network
This commit moves the path conversion helper
routines and tests from `ria_utils.py` to
`support.network`.
