+++ datalad/support/network.py
@@ -6,6 +6,7 @@
 #   copyright and license terms.
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
+from __future__ import annotations
 
 import logging
 
@@ -23,6 +24,7 @@ from hashlib import md5
 from ntpath import splitdrive as win_splitdrive
 from os.path import dirname
 from os.path import join as opj
+from pathlib import PurePosixPath
 from urllib.error import URLError
 from urllib.parse import (
     ParseResult,
@@ -31,6 +33,7 @@ from urllib.parse import (
 from urllib.parse import quote as urlquote
 from urllib.parse import unquote as urlunquote
 from urllib.parse import (
+    quote,
     urlencode,
     urljoin,
     urlparse,
@@ -39,7 +42,6 @@ from urllib.parse import (
 )
 from urllib.request import (
     Request,
-    pathname2url,
     url2pathname,
 )
 
@@ -49,7 +51,6 @@ from datalad import (
     cfg,
     consts,
 )
-from datalad.customremotes.ria_utils import local_path2url_path
 from datalad.support.cache import lru_cache
 from datalad.support.exceptions import CapturedException
 from datalad.utils import (
@@ -1144,6 +1145,62 @@ def download_url(url, dest=None, overwrite=False):
         return providers.download(url, path=str(dest), overwrite=overwrite)
     else:
         return providers.fetch(url)
+
+
+def local_path2url_path(local_path: str,
+                        auto_resolve: bool = False
+                        ) -> str:
+    """Convert a local path into a URL path component"""
+    if not local_path:
+        if not auto_resolve:
+            raise ValueError("cannot convert empty local path to URL path")
+        local_path = os.getcwd()
+
+    url = urlparse(Path(local_path).as_uri())
+    if url.netloc:
+        raise ValueError(
+            f"cannot convert remote path to an URL path: {local_path}")
+    return url.path
+
+
+def url_path2local_path(url_path: str | PurePosixPath,
+                        make_absolute: bool = True,
+                        ) -> str | Path:
+
+    if isinstance(url_path, PurePosixPath):
+        return_path = True
+        url_path = str(url_path)
+    else:
+        return_path = False
+
+    if not url_path or not url_path.startswith("/"):
+        # We expect a 'path-absolute' as defined in RFC 3986, therefore the
+        # path must begin with a slash.
+        raise ValueError(
+            f"url path does not start with '/': {url_path}, and is therefore "
+            f"not an absolute-path as defined in RFC 8089")
+
+    if url_path.startswith("//"):
+        # We expect a 'path-absolute' as defined in RFC 3986, therefore the
+        # first segment must not be empty, i.e. the path must not start with
+        # two or more slashes.
+        raise ValueError(
+            f"url path has empty first segment: {url_path}, and is therefore "
+            f"not an absolute-path as defined in RFC 8089")
+
+    return (
+        Path(url2pathname(url_path))
+        if return_path
+        else url2pathname(url_path)
+    )
+
+
+def quote_path(path: str, safe: str = "/") -> str:
+    """quote the path component of a URL, takes OS specifics into account"""
+    if on_windows:
+        if re.match("^/[a-zA-Z]:/", path):
+            return path[:3] + quote(path[3:], safe=safe)
+    return quote(path, safe=safe)
 
 
 lgr.log(5, "Done importing support.network")
